Is this gonna work on i4i?

+++ aws_workflow.sh
@@ -1,10 +1,31 @@
+INSTANCE_TYPE=$1
+if [ "$INSTANCE_TYPE" != "i4i" ] && [ "$INSTANCE_TYPE" != "x2idn" ]; then
+    echo "Invalid instance type: $INSTANCE_TYPE"
+    exit 1
+fi
+
 # Mount volumes
 echo "Mount volumes: Starting ..."
-DEVICE=/dev/$(lsblk -o NAME,SIZE | grep -E 'nvme[0-9]+n1' | awk '$2 == "16T" {print $1}')
-sudo mkfs -t ext4 $DEVICE
-sudo mkdir /data
-sudo mount $DEVICE /data
-sudo chown $USER:$USER /data
+if [ "$INSTANCE_TYPE" == "i4i" ]; then
+    DEVICE=/dev/nvme8n1
+    sudo mkfs -t ext4 $DEVICE
+    sudo mkdir /data
+    sudo mount $DEVICE /data
+    sudo chown $USER:$USER /data
+    for i in {1..7}; do
+        DEVICE=/dev/nvme${i}n1
+        sudo mkfs -t ext4 $DEVICE
+        sudo mkdir /data-${i}
+        sudo mount $DEVICE /data-${i}
+        sudo chown $USER:$USER /data-${i}
+    done
+elif [ "$INSTANCE_TYPE" == "x2idn" ]; then
+    DEVICE=/dev/$(lsblk -o NAME,SIZE | grep -E 'nvme[0-9]+n1' | awk '$2 == "16T" {print $1}')
+    sudo mkfs -t ext4 $DEVICE
+    sudo mkdir /data
+    sudo mount $DEVICE /data
+    sudo chown $USER:$USER /data
+fi
 echo "Mount volumes: Done"
 echo "================================================"
 
@@ -44,20 +65,31 @@ export INDEX_NAME="v6_${NAME}_u8"
 export MINLEN="200"
 export AWS_MAX_CONCURRENCY=128
 aws s3 sync s3://ai2-llm/pretraining-data/sources/cc_all_dressed/all_dressed_subsamples/deduplication_ablations_v1/minhash_suffarr/minhash_only/ /data/${NAME}/ --only-show-errors
-python indexing_v6_sharded.py \
-    --data_dir /data/${NAME} \
-    --save_dir /data/${INDEX_NAME} \
-    --token_dtype u8 \
-    --cpus 128 \
-    --num_batches 8 \
-    --add_metadata --ulimit 524288
+if [ "$INSTANCE_TYPE" == "i4i" ]; then
+    python indexing_v6_sharded.py \
+        --data_dir /data/${NAME} \
+        --save_dir /data/${INDEX_NAME} \
+        --token_dtype u8 \
+        --cpus 128 \
+        --num_batches 8 \
+        --add_metadata \
+        --split_to_volumes
+elif [ "$INSTANCE_TYPE" == "x2idn" ]; then
+    python indexing_v6_sharded.py \
+        --data_dir /data/${NAME} \
+        --save_dir /data/${INDEX_NAME} \
+        --token_dtype u8 \
+        --cpus 128 \
+        --num_batches 8 \
+        --add_metadata
+fi
 python find_remove_ranges.py \
     --index_dir /data/${INDEX_NAME} \
     --minlen ${MINLEN} \
     --mode parallel_sharded \
     --num_threads 128 \
     --low_ram \
-    --num_batches 2 --ulimit 524288
+    --num_batches 8
 python write_back_to_jsonl_sharded.py \
     --index_dir /data/${INDEX_NAME} \
     --minlen ${MINLEN} \
